FileBasedIndexProjectHandler: use long for file sizes to avoid possible overflow with file >2Gb
intellij-monorepo: f2e133536595d8ffc28331aec69d43be2fa17ad8

+++ intellij-community/platform/lang-impl/src/com/intellij/util/indexing/FileBasedIndexProjectHandler.java
@@ -180,7 +180,7 @@ public class FileBasedIndexProjectHandler implements IndexableFileSet, Disposabl
     long start = System.currentTimeMillis();
     return !index.processChangedFiles(project, new Processor<VirtualFile>() {
       int filesInProjectToBeIndexed;
-      int sizeOfFilesToBeIndexed;
+      long sizeOfFilesToBeIndexed;
 
       @Override
       public boolean process(VirtualFile file) {
